+++ src/enums/mod.rs
@@ -1,20 +1,20 @@
 //! All Enums for Rspotify
-pub mod misc;
 pub mod country;
+pub mod misc;
 pub mod types;
 
 pub use country::Country;
 
-pub use misc::RepeatState;
 pub use misc::DisallowKey;
 pub use misc::IncludeExternal;
+pub use misc::RepeatState;
 pub use misc::TimeRange;
 
 pub use types::AdditionalType;
 pub use types::AlbumType;
 pub use types::CurrentlyPlayingType;
 pub use types::DeviceType;
+pub use types::Error;
 pub use types::ErrorKind;
 pub use types::SearchType;
 pub use types::Type;
-pub use types::Error;