+++ src/model/mod.rs
@@ -25,3 +25,9 @@ pub enum PlayingItem {
     Track(track::FullTrack),
     Episode(show::FullEpisode),
 }
+
+pub use {
+    album::*, artist::*, audio::*, category::*, context::*, cud_result::*, device::*, image::*,
+    offset::*, page::*, playing::*, playlist::*, recommend::*, search::*, show::*, track::*,
+    user::*,
+};