+++ tests/test_with_oauth.rs
@@ -11,9 +11,9 @@ use lazy_static::lazy_static;
 use serde_json::map::Map;
 
 use rspotify::client::Spotify;
-use rspotify::model::offset::for_position;
-use rspotify::oauth2::{SpotifyClientCredentials, SpotifyOAuth};
 use rspotify::enums::{Country, RepeatState, SearchType, TimeRange};
+use rspotify::model::for_position;
+use rspotify::oauth2::{SpotifyClientCredentials, SpotifyOAuth};
 use rspotify::util::get_token;
 
 lazy_static! {