+++ CHANGELOG.md
@@ -44,7 +44,7 @@
 	- `user_follow_users`
 	- `user_unfollow_users`
 	- `audios_features`
-
+- Reexport `model` module to allow user to write `rspotify::model::FullAlbum` instead of  `rspotify::model::FullAlbum`.
 ## 0.10 (2020/07/01)
 
 - Add `get_access_token_without_cache` and `refresh_access_token_without_cache` to get and refresh access token without caching it.